requirements.txt
certifi==2019.3.9
chardet==3.0.4
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11
django-bootstrap3==11.0.0
django-heroku==0.3.1
django-registration==2.4.1
django-tinymce==2.8.0
idna==2.8
Pillow==5.4.1
psycopg2==2.7.7
python-dateutil==2.8.0
python-decouple==3.1
pytz==2018.9
pyuploadcare==2.6.0
requests==2.21.0
six==1.12.0
urllib3==1.24.1
whitenoise==4.1.2
